requirements.txt
django>=4.0.0
openai>=1.0.0
google-generativeai>=0.3.0
gspread>=5.0.0
oauth2client>=4.1.3
python-dotenv>=1.0.0